(PC-35555)[BO] feat: refresh collective offer details page's design
