+++ api/src/pcapi/routes/backoffice/offers/blueprint.py
@@ -6,7 +6,6 @@ import enum
 import functools
 from io import BytesIO
 import logging
-import random
 import re
 import typing
 
@@ -51,12 +50,12 @@ from pcapi.repository import mark_transaction_as_invalid
 from pcapi.repository import on_commit
 from pcapi.repository import repository
 from pcapi.routes.backoffice import utils
+from pcapi.routes.backoffice.connect_as import generate_connect_as_link
 from pcapi.routes.backoffice.filters import format_amount
 from pcapi.routes.backoffice.filters import pluralize
 from pcapi.routes.backoffice.forms import empty as empty_forms
 from pcapi.utils import regions as regions_utils
 from pcapi.utils import string as string_utils
-from pcapi.utils import urls
 
 from . import forms
 
@@ -1179,15 +1178,6 @@ def get_offer_details(offer_id: int) -> utils.BackofficeResponse:
         except offers_exceptions.MoveOfferBaseException:
             pass
 
-    connect_as = None
-    if utils.has_current_user_permission(perm_models.Permissions.CONNECT_AS_PRO):
-        random_int = random.randint(10000, 99999)
-        connect_as_form_id = f"connect-as-form-offer-{offer.id}-{random_int}"
-        pc_pro_url = urls.build_pc_pro_offer_path(offer)
-        connect_as_href = urls.build_pc_pro_offer_link(offer)
-        connect_as_form = forms.ConnectAsForm(object_type="offer", object_id=offer.id, redirect=pc_pro_url)
-        connect_as = {"form": connect_as_form, "form_name": connect_as_form_id, "href": connect_as_href}
-
     return render_template(
         "offer/details_v2.html" if FeatureToggle.WIP_ENABLE_BO_OFFER_DETAILS_V2 else "offer/details.html",
         offer=offer,
@@ -1196,7 +1186,7 @@ def get_offer_details(offer_id: int) -> utils.BackofficeResponse:
         reindex_offer_form=empty_forms.EmptyForm() if is_advanced_pro_support else None,
         edit_offer_venue_form=edit_offer_venue_form,
         move_offer_form=move_offer_form,
-        connect_as=connect_as,
+        connect_as=generate_connect_as_link(offer),
         allowed_actions=allowed_actions,
         action=OfferDetailsActionType,
     )
