+++ api/tests/routes/backoffice/collective_offers_test.py
@@ -27,11 +27,9 @@ from pcapi.models import db
 from pcapi.models.offer_mixin import CollectiveOfferStatus
 from pcapi.models.offer_mixin import OfferValidationStatus
 from pcapi.models.offer_mixin import OfferValidationType
-from pcapi.routes.backoffice.filters import format_date
 from pcapi.utils import db as db_utils
 from pcapi.utils.requests import exceptions as requests_exceptions
 
-from .helpers import button as button_helpers
 from .helpers import html_parser
 from .helpers.get import GetEndpointHelper
 from .helpers.post import PostEndpointHelper
@@ -1356,198 +1354,6 @@ class GetBatchCollectiveOffersRejectFormTest(GetEndpointHelper):
         with assert_num_queries(2):  # session + current user
             response = authenticated_client.get(url)
             assert response.status_code == 200
-
-
-class GetCollectiveOfferDetailTest(GetEndpointHelper):
-    endpoint = "backoffice_web.collective_offer.get_collective_offer_details"
-    endpoint_kwargs = {"collective_offer_id": 1}
-    needed_permission = perm_models.Permissions.READ_OFFERS
-    # Use assert_num_queries() instead of assert_no_duplicated_queries() which does not detect one extra query caused
-    # by a field added in the jinja template.
-    # - fetch session (1 query)
-    # - fetch user (1 query)
-    # - fetch CollectiveOffer
-    # - _is_collective_offer_price_editable
-    expected_num_queries = 4
-    expected_num_queries_with_ff = expected_num_queries + 1  # FF MOVE_OFFER_TEST
-
-    def test_nominal(self, legit_user, authenticated_client):
-        start_date = datetime.datetime.utcnow() - datetime.timedelta(days=1)
-        end_date = start_date + datetime.timedelta(days=28)
-        collective_booking = educational_factories.CollectiveBookingFactory(
-            collectiveStock__startDatetime=start_date,
-            collectiveStock__endDatetime=end_date,
-            collectiveStock__collectiveOffer__teacher=educational_factories.EducationalRedactorFactory(
-                firstName="Pacôme", lastName="De Champignac"
-            ),
-            collectiveStock__collectiveOffer__institution=educational_factories.EducationalInstitutionFactory(
-                name="Ecole de Marcinelle"
-            ),
-            collectiveStock__collectiveOffer__template=educational_factories.CollectiveOfferTemplateFactory(
-                name="offre Vito Cortizone pour lieu que l'on ne peut refuser"
-            ),
-        )
-        url = url_for(self.endpoint, collective_offer_id=collective_booking.collectiveStock.collectiveOffer.id)
-        with assert_num_queries(self.expected_num_queries_with_ff):
-            response = authenticated_client.get(url)
-            assert response.status_code == 200
-
-        content_as_text = html_parser.content_as_text(response.data)
-        assert "Ajuster le prix de l'offre" in content_as_text
-        assert (
-            f"Date de l'évènement : {start_date.strftime('%d/%m/%Y')} → {end_date.strftime('%d/%m/%Y')}"
-            in content_as_text
-        )
-        assert "Statut : Expirée" in content_as_text
-        assert "Statut PC Pro : Réservée" in content_as_text
-        assert "État : • Validée" in content_as_text
-        assert "Utilisateur de la dernière validation" not in content_as_text
-        assert "Date de dernière validation de l’offre" not in content_as_text
-        assert "Enseignant : Pacôme De Champignac" in content_as_text
-        assert "Établissement : Ecole de Marcinelle" in content_as_text
-        assert "Offre vitrine liée : offre Vito Cortizone pour lieu que l'on ne peut refuser" in content_as_text
-
-    def test_processed_pricing(self, legit_user, authenticated_client):
-        pricing = finance_factories.CollectivePricingFactory(
-            status=finance_models.PricingStatus.PROCESSED,
-            collectiveBooking__collectiveStock__startDatetime=datetime.datetime(1970, 1, 1),
-        )
-        url = url_for(self.endpoint, collective_offer_id=pricing.collectiveBooking.collectiveStock.collectiveOffer.id)
-
-        with assert_num_queries(self.expected_num_queries_with_ff):
-            response = authenticated_client.get(url)
-            assert response.status_code == 200
-
-        assert "Ajuster le prix de l'offre" not in response.data.decode()
-
-    def test_invoiced_pricing(self, legit_user, authenticated_client):
-        pricing = finance_factories.CollectivePricingFactory(
-            status=finance_models.PricingStatus.INVOICED,
-            collectiveBooking__collectiveStock__startDatetime=datetime.datetime(1970, 1, 1),
-        )
-        url = url_for(self.endpoint, collective_offer_id=pricing.collectiveBooking.collectiveStock.collectiveOffer.id)
-
-        with assert_num_queries(self.expected_num_queries_with_ff):
-            response = authenticated_client.get(url)
-            assert response.status_code == 200
-
-        assert "Ajuster le prix de l'offre" not in response.data.decode()
-
-    def test_cashflow_pending(self, legit_user, authenticated_client, app):
-        pricing = finance_factories.CollectivePricingFactory(
-            collectiveBooking__collectiveStock__startDatetime=datetime.datetime(1970, 1, 1),
-        )
-        url = url_for(self.endpoint, collective_offer_id=pricing.collectiveBooking.collectiveStock.collectiveOffer.id)
-        app.redis_client.set(finance_conf.REDIS_GENERATE_CASHFLOW_LOCK, "1", 600)
-        try:
-            with assert_num_queries(4):
-                response = authenticated_client.get(url)
-                assert response.status_code == 200
-        finally:
-            app.redis_client.delete(finance_conf.REDIS_GENERATE_CASHFLOW_LOCK)
-
-        assert "Ajuster le prix de l'offre" not in response.data.decode()
-
-    def test_get_validated_offer(self, legit_user, authenticated_client):
-        event_date = datetime.datetime.utcnow() - datetime.timedelta(days=1)
-        validation_date = datetime.datetime.utcnow()
-        collective_booking = educational_factories.CollectiveBookingFactory(
-            collectiveStock__startDatetime=event_date,
-            collectiveStock__collectiveOffer__lastValidationDate=validation_date,
-            collectiveStock__collectiveOffer__validation=offers_models.OfferValidationStatus.APPROVED,
-            collectiveStock__collectiveOffer__lastValidationAuthor=legit_user,
-        )
-        url = url_for(self.endpoint, collective_offer_id=collective_booking.collectiveStock.collectiveOffer.id)
-        with assert_num_queries(self.expected_num_queries_with_ff):
-            response = authenticated_client.get(url)
-
-        content_as_text = html_parser.content_as_text(response.data)
-        assert response.status_code == 200
-        assert f"Utilisateur de la dernière validation : {legit_user.full_name}" in content_as_text
-        assert f"Date de dernière validation : {format_date(validation_date, '%d/%m/%Y à %Hh%M')}" in content_as_text
-
-    def test_get_rejected_offer(self, legit_user, authenticated_client):
-        event_date = datetime.datetime.utcnow() - datetime.timedelta(days=1)
-        validation_date = datetime.datetime.utcnow()
-        collective_booking = educational_factories.CollectiveBookingFactory(
-            collectiveStock__startDatetime=event_date,
-            collectiveStock__collectiveOffer__lastValidationDate=validation_date,
-            collectiveStock__collectiveOffer__validation=offers_models.OfferValidationStatus.REJECTED,
-            collectiveStock__collectiveOffer__lastValidationAuthor=legit_user,
-            collectiveStock__collectiveOffer__rejectionReason=educational_models.CollectiveOfferRejectionReason.MISSING_DESCRIPTION,
-        )
-        url = url_for(self.endpoint, collective_offer_id=collective_booking.collectiveStock.collectiveOffer.id)
-        with assert_num_queries(self.expected_num_queries_with_ff):
-            response = authenticated_client.get(url)
-            assert response.status_code == 200
-
-        content_as_text = html_parser.content_as_text(response.data)
-        assert f"Utilisateur de la dernière validation : {legit_user.full_name}" in content_as_text
-        assert f"Date de dernière validation : {format_date(validation_date, '%d/%m/%Y à %Hh%M')}" in content_as_text
-        assert "Raison de rejet : Description manquante" in content_as_text
-
-    def test_collective_offer_with_offerer_confidence_rule(self, authenticated_client):
-        rule = offerers_factories.ManualReviewOffererConfidenceRuleFactory(offerer__name="Offerer")
-        collective_offer = educational_factories.CollectiveOfferFactory(venue__managingOfferer=rule.offerer)
-
-        url = url_for(self.endpoint, collective_offer_id=collective_offer.id)
-        with assert_num_queries(self.expected_num_queries_with_ff - 1):  # no _is_collective_offer_price_editable
-            response = authenticated_client.get(url)
-            assert response.status_code == 200
-
-        text = html_parser.extract_cards_text(response.data)[0]
-        assert "Entité juridique : Offerer Revue manuelle" in text
-
-    def test_collective_offer_with_venue_confidence_rule(self, authenticated_client):
-        rule = offerers_factories.ManualReviewVenueConfidenceRuleFactory(venue__name="Venue")
-        collective_offer = educational_factories.CollectiveOfferFactory(venue=rule.venue)
-
-        url = url_for(self.endpoint, collective_offer_id=collective_offer.id)
-        with assert_num_queries(self.expected_num_queries_with_ff - 1):  # no _is_collective_offer_price_editable
-            response = authenticated_client.get(url)
-            assert response.status_code == 200
-
-        text = html_parser.extract_cards_text(response.data)[0]
-        assert "Partenaire culturel : Venue Revue manuelle" in text
-
-    def test_collective_offer_with_top_acteur_offerer(self, authenticated_client):
-        collective_offer = educational_factories.CollectiveOfferFactory(
-            venue__managingOfferer__name="Offerer",
-            venue__managingOfferer__tags=[
-                offerers_factories.OffererTagFactory(name="top-acteur", label="Top Acteur"),
-                offerers_factories.OffererTagFactory(name="test", label="Test"),
-            ],
-        )
-
-        url = url_for(self.endpoint, collective_offer_id=collective_offer.id)
-        with assert_num_queries(self.expected_num_queries_with_ff - 1):  # no _is_collective_offer_price_editable
-            response = authenticated_client.get(url)
-            assert response.status_code == 200
-
-        text = html_parser.extract_cards_text(response.data)[0]
-        assert "Entité juridique : Offerer Top Acteur" in text
-
-
-class ValidateCollectiveOfferFromDetailsButtonTest(button_helpers.ButtonHelper):
-    needed_permission = perm_models.Permissions.PRO_FRAUD_ACTIONS
-    button_label = "Valider l'offre"
-    endpoint = "backoffice_web.collective_offer.get_collective_offer_details"
-
-    @property
-    def path(self):
-        stock = educational_factories.CollectiveStockFactory()
-        return url_for(self.endpoint, collective_offer_id=stock.collectiveOffer.id)
-
-
-class RejectCollectiveOfferFromDetailsButtonTest(button_helpers.ButtonHelper):
-    needed_permission = perm_models.Permissions.PRO_FRAUD_ACTIONS
-    button_label = "Rejeter l'offre"
-    endpoint = "backoffice_web.collective_offer.get_collective_offer_details"
-
-    @property
-    def path(self):
-        stock = educational_factories.CollectiveStockFactory()
-        return url_for(self.endpoint, collective_offer_id=stock.collectiveOffer.id)
 
 
 class GetCollectiveOfferPriceFormTest(GetEndpointHelper):
